+++ lib/ui/pages/board/detail_page/widgets/detail_reply_area.dart
@@ -1,0 +1,78 @@
+import 'package:ballkkaye_frontend/_core/style/m_color.dart';
+import 'package:ballkkaye_frontend/_core/style/m_text.dart';
+import 'package:ballkkaye_frontend/ui/pages/board/detail_page/widgets/reply_item.dart';
+import 'package:flutter/material.dart';
+
+class DetailReplyArea extends StatelessWidget {
+  const DetailReplyArea({
+    super.key,
+    required TextEditingController replyController,
+  }) : _replyController = replyController;
+
+  final TextEditingController _replyController;
+
+  @override
+  Widget build(BuildContext context) {
+    final List<Map<String, dynamic>> replies = [
+      {
+        'id': 1,
+        'parentId': null,
+        'author': 'ssar',
+        'content': '댓글 1입니다',
+        'time': '3분 전',
+        'likeCount': 5,
+        'replyCount': 5,
+        'isReplyOwner': false,
+      },
+      {
+        'id': 2,
+        'parentId': 1,
+        'author': 'cos',
+        'content': '대댓글입니다',
+        'time': '2분 전',
+        'likeCount': 5,
+        'replyCount': 5,
+        'isReplyOwner': true,
+      },
+      {
+        'id': 3,
+        'parentId': 1,
+        'author': 'cos',
+        'content': '대댓글입니다',
+        'time': '2분 전',
+        'likeCount': 5,
+        'replyCount': 5,
+        'isReplyOwner': false,
+      },
+      {
+        'id': 4,
+        'parentId': null,
+        'author': 'cos',
+        'content': '대댓글입니다',
+        'time': '2분 전',
+        'likeCount': 5,
+        'replyCount': 5,
+        'isReplyOwner': false,
+      },
+    ]; // TODO : 수정 및 삭제 필요 (참고용)
+
+    return Padding(
+      padding: EdgeInsetsGeometry.symmetric(vertical: 12, horizontal: 16),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          MText.label1_5('댓글 4', color: MColor.kLabel.neutral),
+          ...replies.map((reply) => ReplyItem(
+                reply: reply,
+                onMention: (tagReplyName) {
+                  _replyController.text = tagReplyName;
+                  _replyController.selection = TextSelection.fromPosition(
+                    TextPosition(offset: _replyController.text.length),
+                  );
+                },
+              )), // TODO : 루트 댓글 입력 / 부모가 있는 댓글 입력
+        ],
+      ),
+    );
+  }
+}
